InvoiceProcessing/requirements.txt
amqp==5.2.0
asgiref==3.8.1
asjson==3.0.0
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
celery==5.4.0
certifi==2020.4.5.1
chardet==4.0.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.2.1
coreapi==2.3.3
coreschema==0.0.4
cssselect2==0.7.0
cycler==0.12.1
Django==4.2.13
django-cors-headers==4.4.0
django-filter==2.2.0
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
fonttools==4.53.0
idna==2.9
importlib-metadata==1.6.0
importlib_resources==6.4.0
inflection==0.5.1
isodate==0.6.1
itypes==1.2.0
Jinja2==2.11.2
jsondict==1.2
kiwisolver==1.4.5
kombu==5.3.7
lxml==5.2.2
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.9.0
mrjob==0.7.4
numpy==1.26.4
packaging==24.1
pandas==2.2.2
pillow==10.3.0
platformdirs==4.2.2
Pmw-py3==2.1
preppy==4.2.3
prompt_toolkit==3.0.47
py4j==0.10.9.7
Pygments==2.6.1
PyJWT==1.7.1
pyparsing==3.1.2
pypdf==4.3.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.7
reportlab==4.2.2
requests==2.23.0
requests-file==2.1.0
requests-toolbelt==1.0.0
rest-framework-simplejwt==0.0.2
rlextra==4.2.2
six==1.16.0
sqlparse==0.3.1
svglib==1.4.1
tinycss2==1.3.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==3.0.1
urllib3==1.25.9
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
zeep==4.2.1
zipp==3.1.0
